custom_mmio: Extract MmioOps trait into backend module
Move the MmioOps trait definition and its provided read/write
implementations into a new backend::mmio_ops module, so they can be
reused by the other backends.

No functional change.
