+++ src/backend/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2026 The safe-mmio Authors.
+// This project is dual-licensed under Apache 2.0 and MIT terms.
+// See LICENSE-APACHE and LICENSE-MIT for details.
+
+#[cfg(feature = "custom-mmio")]
+pub mod mmio_ops;
